add bin/compile and bin/detect
The `bin/detect` does nothing but print the buildpack name. The
bin/compile is the real deal.

It fetches the source code, compiles and caches it. This will save a lot
of time on the next builds. Then it installs from the cache. Notice it
will produce only shared libs, as it is what we need.